cplusplus/conan/pyproject.toml
[project]
name = "learn-conan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "conan>=2.11.0",
]
